unofficial version 2.1.2
Bind: temperature, humidity, soil moisture, battery. Setting up reports - works. Fixed the problem of the device running without sleep when switching to a new router. The mode of operation of the button when connecting and leaving the network has been changed. To enter the network, press and hold the button for 5-7 seconds. To exit the network, press and hold the button for 10 seconds, at the end the LED will light up, the button can be released.